+++ src/result/CoverageResult.php
@@ -12,7 +12,6 @@
 namespace cloak\result;
 
 use cloak\value\Coverage;
-use cloak\CoverageResultVisitorInterface;
 
 
 /**